Move metadata tests from descriptor
- Also remove the "# noqa: ANN401"
